backend/requirements.txt
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asyncio==3.4.3
attrs==25.3.0
aws_lambda_powertools==3.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
black==25.1.0
boto3==1.38.27
boto3-stubs==1.39.14
botocore==1.38.27
botocore-stubs==1.38.46
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
dataclasses-json==0.6.7
datamodel-code-generator==0.32.0
distro==1.9.0
emoji==2.14.1
eval_type_backport==0.2.2
fastapi==0.116.1
filetype==1.2.0
frozenlist==1.6.0
genson==1.3.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-experimental==0.3.4
langchain-openai==0.3.28
langchain-pinecone==0.2.11
langchain-tests==0.3.20
langchain-text-splitters==0.3.9
langdetect==1.0.9
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-checkpoint-amazon-dynamodb==0.1.2
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.4.10
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mangum==0.8.0
mdurl==0.1.2
more-itertools==10.7.0
moto==5.1.9
multidict==6.4.3
mypy-boto3-bedrock==1.39.12
mypy-boto3-bedrock-agent==1.39.5
mypy-boto3-bedrock-agent-runtime==1.39.0
mypy-boto3-bedrock-runtime==1.39.7
mypy-boto3-lambda==1.39.11
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
olefile==0.47
ollama==0.4.8
openai==1.98.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pathspec==0.12.1
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.2
Pympler==1.1
pypdf==5.4.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==4.0.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.25.7
rich==14.1.0
s3transfer==0.13.1
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.47.2
syrupy==4.9.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typeguard==4.4.4
types-awscrt==0.27.4
types-s3transfer==0.13.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions
unstructured==0.17.2
unstructured-client==0.34.0
urllib3==2.4.0
uvicorn==0.35.0
uvloop==0.21.0; sys_platform != "win32" # remove the ; sys_platform != "win32" line of code if you encounter issues intsalling it for macos 
vcrpy==7.0.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
